增加Lucene搜索集成测试和数据填充集成测试

+++ src/main/java/com/timeyang/amanda/base/search/SearchResult.java
@@ -15,5 +15,5 @@ import lombok.Setter;
 @AllArgsConstructor
 public class SearchResult<T> {
     private final T entity;
-    private final double relevance;
+    private final Float relevance;
 }
